% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corrplot.R
\name{Corrplot}
\alias{Corrplot}
\title{Plotting corrplot based on input TIMEscore and specified gene}
\usage{
Corrplot(
  score,
  expr,
  gene,
  method = "pearson",
  color = "#C45A44",
  width = 4.5,
  height = 4.5,
  filename = "corrplot"
)
}
\arguments{
\item{score}{a data.frame which is the output of function 'TIMEscore'}

\item{expr}{a data.frame rownamed with gene symbols and colnamed with sample IDs}

\item{gene}{gene symbol, for example, CD274}

\item{method}{method performing correlation analysis, one of 'pearson' and 'spearman'}

\item{color}{color of corrplot}

\item{width}{width of pdf file}

\item{height}{height of pdf file}

\item{filename}{output filename}
}
\value{
a pdf file of corrplot
}
\description{
Plotting corrplot based on input TIMEscore and specified gene
}
\examples{

data(exprSet)
exprSet[1:5, 1:5]
score <- TIMEscore(expr = exprSet)
Corrplot(score, exprSet, "CD274")

}
