% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprSet}
\alias{exprSet}
\title{a example expression matrix which is a data.frame rownamed with gene symbols and colnamed with sample IDs}
\format{
a data.frame
}
\usage{
exprSet
}
\description{
a example expression matrix which is a data.frame rownamed with gene symbols and colnamed with sample IDs
}
\keyword{datasets}
