% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCA_df_n}
\alias{PCA_df_n}
\title{a data.frame containing the protected genes with their coefficients}
\format{
a data.frame
}
\usage{
PCA_df_n
}
\description{
a data.frame containing the protected genes with their coefficients
}
\keyword{datasets}
