% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCA_df_p}
\alias{PCA_df_p}
\title{a data.frame containing the hazardous genes with their coefficients}
\format{
a data.frame
}
\usage{
PCA_df_p
}
\description{
a data.frame containing the hazardous genes with their coefficients
}
\keyword{datasets}
