% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival}
\alias{survival}
\title{example survival information which is a data.frame rownamed with sample IDs identical to exprSet}
\format{
a data.frame
}
\usage{
survival
}
\description{
example survival information which is a data.frame rownamed with sample IDs identical to exprSet
}
\keyword{datasets}
