% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survplot.R
\name{Survplot}
\alias{Survplot}
\title{Plotting survival curve based on input TIMEscore and survival information}
\usage{
Survplot(
  score,
  survival,
  unit = "Days",
  color = c("#cf273c", "#5091c0"),
  linesize = 1,
  width = 4.5,
  height = 5.5,
  filename = "survplot"
)
}
\arguments{
\item{score}{a data.frame which is the output of function 'TIMEscore'}

\item{survival}{survival information which is a data.frame rownamed with sample IDs identical to the rownames of output of function 'TIMEscore'. It is colnamed with 'status' and 'time' separately.}

\item{unit}{time unit of time column in param survival, which is one of ’Days‘, 'Months', 'Years'.}

\item{color}{color vector of high- and low-group}

\item{linesize}{a number referring to the line size of high- and low-group}

\item{width}{width of pdf file}

\item{height}{height of pdf file}

\item{filename}{output filename}
}
\value{
a pdf file of survival curve
}
\description{
Plotting survival curve based on input TIMEscore and survival information
}
\examples{

data(exprSet)
exprSet[1:5, 1:5]
data(survival)
head(survival)
score <- TIMEscore(expr = exprSet)
Survplot(score, survival)

}
