% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeROC.R
\name{TimeROC}
\alias{TimeROC}
\title{Plotting time-dependent ROC based on input TIMEscore and survival information}
\usage{
TimeROC(
  score,
  survival,
  times = c(1, 2, 3),
  color = c("slateblue", "seagreen3", "firebrick1"),
  title = "Time-dependent ROC",
  width = 4,
  height = 4,
  filename = "timeROC"
)
}
\arguments{
\item{score}{a data.frame which is the output of function 'TIMEscore'}

\item{survival}{survival information which is a data.frame rownamed with sample IDs identical to the rownames of output of function 'TIMEscore'. It is colnamed with 'status' and 'time' separately. Time unit has to be 'Year'.}

\item{times}{a numeric vector, time points for calculating AUC}

\item{color}{color vector of different time-dependent ROC}

\item{title}{main title of timeROC}

\item{width}{width of pdf file}

\item{height}{height of pdf file}

\item{filename}{output filename}
}
\value{
a pdf file of time-dependent ROC
}
\description{
Plotting time-dependent ROC based on input TIMEscore and survival information
}
\examples{

data(exprSet)
exprSet[1:5, 1:5]
data(survival)
survival2 <- survival
survival2$time <- survival2$time / 365
head(survival2)
score <- TIMEscore(expr = exprSet)
library(survival)
TimeROC(score, survival2)

}
