% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMEscore.R
\name{TIMEscore}
\alias{TIMEscore}
\title{Calculating TIMEscore for every sample.}
\usage{
TIMEscore(expr)
}
\arguments{
\item{expr}{a data.frame rownamed with gene symbols and colnamed with sample IDs}
}
\value{
a data.frame containing sample IDs and corresponding TIMEscore
}
\description{
Calculating TIMEscore for every sample.
}
\examples{

data(exprSet)
exprSet[1:5, 1:5]
result <- TIMEscore(expr = exprSet)
head(result)
}
